# src/gen/doc/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the sample swagger file for appdirect code challenge",
    "version" : "1.0.0",
    "title" : "Appdirect Challenge Endpoints",
    "contact" : {
      "email" : "krunal_patel9445@yahoo.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "petstore.swagger.io",
  "basePath" : "/api",
  "tags" : [ {
    "name" : "Subscription"
  }, {
    "name" : "User"
  } ],
  "schemes" : [ "https" ],
  "paths" : {
    "/subscription/create" : {
      "get" : {
        "tags" : [ "Subscription" ],
        "summary" : "Create a new subscription",
        "description" : "",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "eventUrl",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    },
    "/subscription/cancel" : {
      "get" : {
        "tags" : [ "Subscription" ],
        "summary" : "Cancel a subscription",
        "description" : "",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "eventUrl",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    },
    "/subscription/change" : {
      "get" : {
        "tags" : [ "Subscription" ],
        "summary" : "Modify a subscription",
        "description" : "",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "eventUrl",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    },
    "/subscription/notification" : {
      "get" : {
        "tags" : [ "Subscription" ],
        "summary" : "Notification of the subscription",
        "description" : "",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "eventUrl",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    },
    "/user/assignUser" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "Assign user(s) to the subscription",
        "description" : "",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "eventUrl",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    },
    "/user/unassignUser" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "Unassign user(s) from the subscription",
        "description" : "",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "eventUrl",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Response"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Subscription" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string",
          "enum" : [ "SUBSCRIPTION_ORDER", "SUBSCRIPTION_CANCEL" ]
        },
        "marketPlace" : {
          "$ref" : "#/MarketPlace"
        },
        "applicationUuid" : {
          "type" : "string"
        },
        "flag" : {
          "type" : "string",
          "enum" : [ "STATELESS", "DEVELOPMENT" ]
        },
        "returnUrl" : {
          "type" : "string"
        },
        "creator" : {
          "$ref" : "#/User"
        },
        "payload" : {
          "$ref" : "#/Payload"
        }
      }
    },
    "MarketPlace" : {
      "type" : "object",
      "properties" : {
        "baseUrl" : {
          "type" : "string"
        },
        "partner" : {
          "type" : "string"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "openId" : {
          "type" : "string"
        },
        "uuid" : {
          "type" : "string"
        },
        "locale" : {
          "type" : "string"
        },
        "address" : {
          "$ref" : "#/Address"
        }
      }
    },
    "Address" : {
      "type" : "object",
      "properties" : {
        "city" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        },
        "state" : {
          "type" : "string"
        },
        "street1" : {
          "type" : "string"
        },
        "street2" : {
          "type" : "string"
        },
        "zip" : {
          "type" : "string"
        }
      }
    },
    "Payload" : {
      "type" : "object",
      "properties" : {
        "user" : {
          "$ref" : "#/User"
        },
        "account" : {
          "$ref" : "#/Account"
        },
        "company" : {
          "$ref" : "#/Company"
        },
        "order" : {
          "$ref" : "#/Order"
        },
        "addonInstance" : {
          "$ref" : "#/AddonInstance"
        },
        "addonBinding" : {
          "$ref" : "#/AddonBinding"
        },
        "notice" : {
          "$ref" : "#/Notice"
        }
      }
    },
    "Account" : {
      "type" : "object",
      "properties" : {
        "accountIdentifier" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        }
      }
    },
    "Company" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string"
        },
        "externalId" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "website" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        }
      }
    },
    "Order" : {
      "type" : "object",
      "properties" : {
        "editionCode" : {
          "type" : "string"
        },
        "pricingDuration" : {
          "type" : "string"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Order_items"
          }
        }
      }
    },
    "AddonInstance" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        }
      }
    },
    "AddonBinding" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        }
      }
    },
    "Notice" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string",
          "enum" : [ "REACTIVATED", "DEACTIVATED", "CLOSED" ]
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Response" : {
      "type" : "object",
      "properties" : {
        "success" : {
          "type" : "boolean"
        },
        "accountIdentifier" : {
          "type" : "string"
        },
        "errorCode" : {
          "type" : "string",
          "enum" : [ "USER_ALREADY_EXISTS", "USER_NOT_FOUND", "ACCOUNT_NOT_FOUND", "MAX_USERS_REACHED", "UNAUTHORIZED", "OPERATION_CANCELED", "CONFIGURATION_ERROR", "INVALID_RESPONSE", "PENDING", "FORBIDDEN", "BINDING_NOT_FOUND", "TRANSPORT_ERROR", "UNKNOWN_ERROR" ]
        },
        "message" : {
          "type" : "string"
        },
        "userIdentifier" : {
          "type" : "string"
        }
      }
    },
    "Order_items" : {
      "properties" : {
        "quantity" : {
          "type" : "integer"
        },
        "unit" : {
          "type" : "string"
        }
      }
    }
  }
}